Separated services for better organization and clarity, removed unused methods/services, php doc comment cleanup, removed redundant parameters/default values
