+++ src/Controller/Studio/ClassificationStore/LoadAttributesController.php
@@ -45,9 +45,6 @@ final class LoadAttributesController extends AbstractApiController
         parent::__construct($serializer);
     }
 
-    /**
-     * @throws \Exception
-     */
     #[Route(
         path: self::ROUTE,
         name: 'pimcore_studio_api_data_importer_classificationstore_load_attributes',
