+++ src/Controller/Studio/ClassificationStore/LoadKeyNameController.php
@@ -45,9 +45,6 @@ final class LoadKeyNameController extends AbstractApiController
         parent::__construct($serializer);
     }
 
-    /**
-     * @throws \Exception
-     */
     #[Route(
         path: self::ROUTE,
         name: 'pimcore_studio_api_data_importer_classificationstore_load_key_name',
