+++ src/Controller/Studio/Config/CopyPreviewController.php
@@ -18,7 +18,7 @@ use OpenApi\Attributes\Schema;
 use Pimcore\Bundle\DataImporterBundle\OpenApi\Config\Prefix;
 use Pimcore\Bundle\DataImporterBundle\OpenApi\Config\Tags;
 use Pimcore\Bundle\DataImporterBundle\Schema\CopyPreviewParameters;
-use Pimcore\Bundle\DataImporterBundle\Service\Studio\ConfigurationServiceInterface;
+use Pimcore\Bundle\DataImporterBundle\Service\Studio\PreviewDataServiceInterface;
 use Pimcore\Bundle\DataImporterBundle\Utils\Constants\PermissionConstants;
 use Pimcore\Bundle\StudioBackendBundle\Controller\AbstractApiController;
 use Pimcore\Bundle\StudioBackendBundle\OpenApi\Attribute\Parameter\Path\IdParameter;
@@ -41,14 +41,11 @@ final class CopyPreviewController extends AbstractApiController
 
     public function __construct(
         SerializerInterface $serializer,
-        private readonly ConfigurationServiceInterface $configurationService
+        private readonly PreviewDataServiceInterface $previewDataService
     ) {
         parent::__construct($serializer);
     }
 
-    /**
-     * @throws \Exception
-     */
     #[Route(
         path: self::ROUTE,
         name: 'pimcore_studio_api_data_importer_config_copy_preview',
@@ -81,7 +78,7 @@ final class CopyPreviewController extends AbstractApiController
         string $name,
         #[MapRequestPayload] CopyPreviewParameters $parameters
     ): Response {
-        $this->configurationService->copyPreviewData($name, $parameters->getCurrentConfig());
+        $this->previewDataService->copyPreviewData($name, $parameters->getCurrentConfig());
 
         return new Response();
     }
