+++ src/Controller/Studio/Config/GetController.php
@@ -45,9 +45,6 @@ final class GetController extends AbstractApiController
         parent::__construct($serializer);
     }
 
-    /**
-     * @throws \Exception
-     */
     #[Route(
         path: self::ROUTE,
         name: 'pimcore_studio_api_data_importer_config_get',
