+++ src/Controller/Studio/Config/HasImportFileUploadedController.php
@@ -19,7 +19,7 @@ use OpenApi\Attributes\Schema;
 use Pimcore\Bundle\DataImporterBundle\OpenApi\Config\Prefix;
 use Pimcore\Bundle\DataImporterBundle\OpenApi\Config\Tags;
 use Pimcore\Bundle\DataImporterBundle\Schema\ImportFileStatusResponse;
-use Pimcore\Bundle\DataImporterBundle\Service\Studio\ConfigurationServiceInterface;
+use Pimcore\Bundle\DataImporterBundle\Service\Studio\ImportServiceInterface;
 use Pimcore\Bundle\DataImporterBundle\Utils\Constants\PermissionConstants;
 use Pimcore\Bundle\StudioBackendBundle\Controller\AbstractApiController;
 use Pimcore\Bundle\StudioBackendBundle\OpenApi\Attribute\Parameter\Path\IdParameter;
@@ -40,14 +40,11 @@ final class HasImportFileUploadedController extends AbstractApiController
 
     public function __construct(
         SerializerInterface $serializer,
-        private readonly ConfigurationServiceInterface $configurationService
+        private readonly ImportServiceInterface $importService
     ) {
         parent::__construct($serializer);
     }
 
-    /**
-     * @throws \Exception
-     */
     #[Route(
         path: self::ROUTE,
         name: 'pimcore_studio_api_data_importer_config_has_import_file_uploaded',
@@ -78,7 +75,7 @@ final class HasImportFileUploadedController extends AbstractApiController
     public function hasImportFileUploaded(string $name): JsonResponse
     {
         return $this->jsonResponse(
-            $this->configurationService->hasImportFileUploaded($name)
+            $this->importService->hasImportFileUploaded($name)
         );
     }
 }
