+++ src/Controller/Studio/Config/LoadPreviewController.php
@@ -20,7 +20,7 @@ use Pimcore\Bundle\DataImporterBundle\OpenApi\Config\Prefix;
 use Pimcore\Bundle\DataImporterBundle\OpenApi\Config\Tags;
 use Pimcore\Bundle\DataImporterBundle\Schema\DataPreviewResponse;
 use Pimcore\Bundle\DataImporterBundle\Schema\LoadPreviewParameters;
-use Pimcore\Bundle\DataImporterBundle\Service\Studio\ConfigurationServiceInterface;
+use Pimcore\Bundle\DataImporterBundle\Service\Studio\PreviewDataServiceInterface;
 use Pimcore\Bundle\DataImporterBundle\Utils\Constants\PermissionConstants;
 use Pimcore\Bundle\StudioBackendBundle\Controller\AbstractApiController;
 use Pimcore\Bundle\StudioBackendBundle\OpenApi\Attribute\Parameter\Path\IdParameter;
@@ -43,14 +43,11 @@ final class LoadPreviewController extends AbstractApiController
 
     public function __construct(
         SerializerInterface $serializer,
-        private readonly ConfigurationServiceInterface $configurationService
+        private readonly PreviewDataServiceInterface $previewDataService
     ) {
         parent::__construct($serializer);
     }
 
-    /**
-     * @throws \Exception
-     */
     #[Route(
         path: self::ROUTE,
         name: 'pimcore_studio_api_data_importer_config_load_preview',
@@ -84,7 +81,7 @@ final class LoadPreviewController extends AbstractApiController
         #[MapRequestPayload] LoadPreviewParameters $parameters
     ): JsonResponse {
         return $this->jsonResponse(
-            $this->configurationService->loadPreviewData(
+            $this->previewDataService->loadPreviewData(
                 $name,
                 $parameters->getCurrentConfig(),
                 $parameters->getRecordNumber()
