+++ src/Controller/Studio/Config/LoadTransformationResultController.php
@@ -20,7 +20,7 @@ use Pimcore\Bundle\DataImporterBundle\OpenApi\Config\Prefix;
 use Pimcore\Bundle\DataImporterBundle\OpenApi\Config\Tags;
 use Pimcore\Bundle\DataImporterBundle\Schema\LoadPreviewParameters;
 use Pimcore\Bundle\DataImporterBundle\Schema\TransformationResultPreviewsResponse;
-use Pimcore\Bundle\DataImporterBundle\Service\Studio\ConfigurationServiceInterface;
+use Pimcore\Bundle\DataImporterBundle\Service\Studio\TransformationServiceInterface;
 use Pimcore\Bundle\DataImporterBundle\Utils\Constants\PermissionConstants;
 use Pimcore\Bundle\StudioBackendBundle\Controller\AbstractApiController;
 use Pimcore\Bundle\StudioBackendBundle\OpenApi\Attribute\Parameter\Path\IdParameter;
@@ -43,14 +43,11 @@ final class LoadTransformationResultController extends AbstractApiController
 
     public function __construct(
         SerializerInterface $serializer,
-        private readonly ConfigurationServiceInterface $configurationService
+        private readonly TransformationServiceInterface $transformationService
     ) {
         parent::__construct($serializer);
     }
 
-    /**
-     * @throws \Exception
-     */
     #[Route(
         path: self::ROUTE,
         name: 'pimcore_studio_api_data_importer_config_load_transformation_result',
@@ -84,7 +81,7 @@ final class LoadTransformationResultController extends AbstractApiController
         #[MapRequestPayload] LoadPreviewParameters $parameters
     ): JsonResponse {
         return $this->jsonResponse(
-            $this->configurationService->loadTransformationResultPreviews(
+            $this->transformationService->loadTransformationResultPreviews(
                 $name,
                 $parameters->getCurrentConfig(),
                 $parameters->getRecordNumber()
