+++ src/Controller/Studio/Config/SaveController.php
@@ -48,9 +48,6 @@ final class SaveController extends AbstractApiController
         parent::__construct($serializer);
     }
 
-    /**
-     * @throws \Exception
-     */
     #[Route(
         path: self::ROUTE,
         name: 'pimcore_studio_api_data_importer_config_save',
