+++ src/Controller/Studio/Config/UploadImportFileController.php
@@ -18,7 +18,7 @@ use OpenApi\Attributes\Property;
 use OpenApi\Attributes\Schema;
 use Pimcore\Bundle\DataImporterBundle\OpenApi\Config\Prefix;
 use Pimcore\Bundle\DataImporterBundle\OpenApi\Config\Tags;
-use Pimcore\Bundle\DataImporterBundle\Service\Studio\ConfigurationServiceInterface;
+use Pimcore\Bundle\DataImporterBundle\Service\Studio\ImportServiceInterface;
 use Pimcore\Bundle\DataImporterBundle\Utils\Constants\PermissionConstants;
 use Pimcore\Bundle\StudioBackendBundle\Controller\AbstractApiController;
 use Pimcore\Bundle\StudioBackendBundle\Exception\Api\EnvironmentException;
@@ -43,14 +43,11 @@ final class UploadImportFileController extends AbstractApiController
 
     public function __construct(
         SerializerInterface $serializer,
-        private readonly ConfigurationServiceInterface $configurationService
+        private readonly ImportServiceInterface $importService
     ) {
         parent::__construct($serializer);
     }
 
-    /**
-     * @throws \Exception
-     */
     #[Route(
         path: self::ROUTE,
         name: 'pimcore_studio_api_data_importer_config_upload_import_file',
@@ -97,7 +94,7 @@ final class UploadImportFileController extends AbstractApiController
             throw new EnvironmentException('Invalid file found in the request');
         }
 
-        $this->configurationService->uploadImportFile($name, $file);
+        $this->importService->uploadImportFile($name, $file);
 
         return new Response();
     }
