+++ src/Service/Studio/ConfigurationServiceInterface.php
@@ -13,16 +13,10 @@ declare(strict_types=1);
 
 namespace Pimcore\Bundle\DataImporterBundle\Service\Studio;
 
-use Pimcore\Bundle\DataImporterBundle\Schema\ColumnHeadersResponse;
 use Pimcore\Bundle\DataImporterBundle\Schema\ConfigurationDetail;
-use Pimcore\Bundle\DataImporterBundle\Schema\CronValidationResponse;
-use Pimcore\Bundle\DataImporterBundle\Schema\DataPreviewResponse;
-use Pimcore\Bundle\DataImporterBundle\Schema\ImportFileStatusResponse;
-use Pimcore\Bundle\DataImporterBundle\Schema\ImportProgressResponse;
-use Pimcore\Bundle\DataImporterBundle\Schema\ImportStartResponse;
-use Pimcore\Bundle\DataImporterBundle\Schema\TransformationResultPreviewsResponse;
-use Pimcore\Bundle\DataImporterBundle\Schema\TransformationResultTypeResponse;
-use Symfony\Component\HttpFoundation\File\UploadedFile;
+use Pimcore\Bundle\StudioBackendBundle\Exception\Api\ConflictException;
+use Pimcore\Bundle\StudioBackendBundle\Exception\Api\ForbiddenException;
+use Symfony\Component\HttpKernel\Exception\NotFoundHttpException;
 
 /**
  * @internal
@@ -30,69 +24,15 @@ use Symfony\Component\HttpFoundation\File\UploadedFile;
 interface ConfigurationServiceInterface
 {
     /**
-     * @throws \Exception
+     * @throws NotFoundHttpException
+     * @throws ForbiddenException
      */
     public function getConfiguration(string $name): ConfigurationDetail;
 
     /**
-     * @throws \Exception
+     * @throws NotFoundHttpException
+     * @throws ForbiddenException
+     * @throws ConflictException
      */
     public function saveConfiguration(string $name, array $configuration, int $modificationDate): ConfigurationDetail;
-
-    /**
-     * @throws \Exception
-     */
-    public function uploadPreviewData(string $name, UploadedFile $file): void;
-
-    /**
-     * @throws \Exception
-     */
-    public function copyPreviewData(string $name, ?array $currentConfig): void;
-
-    /**
-     * @throws \Exception
-     */
-    public function loadPreviewData(string $name, ?array $currentConfig, int $recordNumber): DataPreviewResponse;
-
-    /**
-     * @throws \Exception
-     */
-    public function loadColumnHeaders(string $name, ?array $currentConfig): ColumnHeadersResponse;
-
-    /**
-     * @throws \Exception
-     */
-    public function loadTransformationResultPreviews(string $name, ?array $currentConfig, int $recordNumber): TransformationResultPreviewsResponse;
-
-    /**
-     * @throws \Exception
-     */
-    public function calculateTransformationResultType(string $name, array $currentConfig): TransformationResultTypeResponse;
-
-    /**
-     * @throws \Exception
-     */
-    public function startImport(string $name): ImportStartResponse;
-
-    /**
-     * @throws \Exception
-     */
-    public function checkImportProgress(string $name): ImportProgressResponse;
-
-    /**
-     * @throws \Exception
-     */
-    public function cancelExecution(string $name): void;
-
-    public function validateCronExpression(string $cronExpression): CronValidationResponse;
-
-    /**
-     * @throws \Exception
-     */
-    public function uploadImportFile(string $name, UploadedFile $file): void;
-
-    /**
-     * @throws \Exception
-     */
-    public function hasImportFileUploaded(string $name): ImportFileStatusResponse;
 }
