+++ src/Service/Studio/TransformationServiceInterface.php
@@ -1,0 +1,49 @@
+<?php
+declare(strict_types=1);
+
+/**
+ * This source file is available under the terms of the
+ * Pimcore Open Core License (POCL)
+ * Full copyright and license information is available in
+ * LICENSE.md which is distributed with this source code.
+ *
+ *  @copyright  Copyright (c) Pimcore GmbH (https://www.pimcore.com)
+ *  @license    Pimcore Open Core License (POCL)
+ */
+
+namespace Pimcore\Bundle\DataImporterBundle\Service\Studio;
+
+use Pimcore\Bundle\DataImporterBundle\Exception\InvalidConfigurationException;
+use Pimcore\Bundle\DataImporterBundle\Schema\TransformationResultPreviewsResponse;
+use Pimcore\Bundle\DataImporterBundle\Schema\TransformationResultTypeResponse;
+use Pimcore\Bundle\StudioBackendBundle\Exception\Api\EnvironmentException;
+use Pimcore\Bundle\StudioBackendBundle\Exception\Api\ForbiddenException;
+use Symfony\Component\HttpKernel\Exception\NotFoundHttpException;
+
+/**
+ * @internal
+ */
+interface TransformationServiceInterface
+{
+    /**
+     * @throws NotFoundHttpException
+     * @throws ForbiddenException
+     * @throws EnvironmentException
+     * @throws InvalidConfigurationException
+     */
+    public function loadTransformationResultPreviews(
+        string $name,
+        ?array $currentConfig,
+        int $recordNumber
+    ): TransformationResultPreviewsResponse;
+
+    /**
+     * @throws NotFoundHttpException
+     * @throws ForbiddenException
+     * @throws InvalidConfigurationException
+     */
+    public function calculateTransformationResultType(
+        string $name,
+        array $currentConfig
+    ): TransformationResultTypeResponse;
+}
